+++ src/Sulu/Bundle/AutomationBundle/Events/TaskUpdateEvent.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\AutomationBundle\Events;
+
+/**
+ * Event args for task-update event.
+ */
+class TaskUpdateEvent extends TaskEvent
+{
+    use CancelTrait;
+}