+++ src/Sulu/Bundle/ContentBundle/DependencyInjection/SuluContentExtension.php
@@ -131,7 +131,7 @@ class SuluContentExtension extends Extension implements PrependExtensionInterfac
 
         $this->processTemplates($container, $config);
 
-        if (isset($bundles['SuluSearchBundle'])) {
+        if (array_key_exists('SuluSearchBundle', $bundles)) {
             $this->processSearch($config, $loader, $container);
         }
 
@@ -148,6 +148,10 @@ class SuluContentExtension extends Extension implements PrependExtensionInterfac
         $loader->load('export.xml');
         $loader->load('command.xml');
         $loader->load('link-tag.xml');
+
+        if (array_key_exists('SuluAutomationBundle', $bundles)) {
+            $loader->load('automation.xml');
+        }
     }
 
     private function processTemplates(ContainerBuilder $container, $config)