+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/main.js
@@ -976,6 +976,7 @@ define([
                     tabs: {
                         url: navigationUrl,
                         options: {
+                            locale: this.options.language,
                             data: function() {
                                 return this.sandbox.util.deepCopy(this.content.toJSON());
                             }.bind(this)