+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG for Sulu
     * ENHANCEMENT #3059 [ContentBundle]       Fixed deprecation in WebspaceCollection rendering
     * ENHANCEMENT #3058 [All]                 Fixed some config files for symfony3
     * BUGFIX      #3043 [ContentBudle]        Fixed bind null values on managed-structure
+    * FEATURE     #3037 [AutomationBundle]    Added automation-bundle
     * ENHANCEMENT #2856 [ContactBundle]       Removed not needed css
     * BUGFIX      #3034 [LocationBundle]      Load external map data over https
     * BUGFIX      #3031 [AdminBundle]         Fixed defaultDisplayOption in media selectio content type